+++ app/models/cart_item/space_reservation.rb
@@ -13,6 +13,8 @@ class CartItem::SpaceReservation < CartItem::Reservation
 
   belongs_to :plan
 
+  belongs_to :reservation_context
+
   def type
     'space'
   end