+++ db/migrate/20230718134350_add_reservation_context_id_to_reservation.rb
@@ -1,0 +1,5 @@
+class AddReservationContextIdToReservation < ActiveRecord::Migration[7.0]
+  def change
+    add_reference :reservations, :reservation_context, index: true, foreign_key: true
+  end
+end