+++ test/frontend/__fixtures__/settings.ts
@@ -855,6 +855,12 @@ export const settings: Array<Setting> = [
     value: 'false',
     last_update: '2023-03-31T14:39:12+0100',
     localized: 'Family account'
+  },
+  {
+    name: 'reservation_context_feature',
+    value: 'false',
+    last_update: '2022-12-23T14:39:12+0100',
+    localized: 'Reservation context feature'
   }
 ];